% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{adp_download}
\alias{adp_download}
\title{Download adp and change name of files}
\usage{
adp_download(path)
}
\arguments{
\item{path}{path for downloading and unziping files}
}
\description{
Function for downloading and renaming adp data
}
