% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicer.R
\name{adp_slicer}
\alias{adp_slicer}
\title{Slicer}
\usage{
adp_slicer(file)
}
\arguments{
\item{file}{Shapefile to slice}
}
\description{
Slices sf object of subdivision to the smallest
possible elements by adp code
}
