% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adp.R
\name{get_adp}
\alias{get_adp}
\title{Getting sf object of adp}
\usage{
get_adp(adp_type, name = NULL, code = NULL)
}
\arguments{
\item{adp_type}{one of subdivision levels (panstwo, wojewodztwa, powiaty, gminy,
jednostki ewidencyjne, obreby ewidencyjne)}

\item{name}{name of subdivision}

\item{code}{code of subdivision}
}
\value{
sf object
}
\description{
Returns sf object of chosen subdivision/subdivisions
}
