% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_adp.R
\name{info_adp}
\alias{info_adp}
\title{Getting names and codes of subdivisions}
\usage{
info_adp(adp_type)
}
\arguments{
\item{adp_type}{one of subdivision levels (panstwo, wojewodztwa, powiaty, gminy,
jednostki ewidencyjne, obreby ewidencyjne)}
}
\value{
data frame class(base)
}
\description{
Returns data frame with names and codes of chosen subdivision
}
